Refactor login modal store location and imports
Moved the login modal Zustand store from components/Login/store.jsx to stores/login-modal.jsx for better organization. Updated all relevant imports in LoginModal, InputEditor Table, and Tool components to use the new store path.
