+++ src/features/auth/stores/login-modal.jsx
@@ -1,12 +1,12 @@
 import { create } from 'zustand';
 
-export const useLoginStore = create((set) => ({
+const useLoginModalStore = create((set) => ({
   showLoginModal: false,
   setShowLoginModal: (showLoginModal) => set({ showLoginModal }),
 }));
 
 export const useShowLoginModal = () =>
-  useLoginStore((state) => state.showLoginModal);
+  useLoginModalStore((state) => state.showLoginModal);
 
 export const useSetShowLoginModal = () =>
-  useLoginStore((state) => state.setShowLoginModal);
+  useLoginModalStore((state) => state.setShowLoginModal);
